# Improve Shapes layer performance