--- a/napari/layers/base/base.py
+++ b/napari/layers/base/base.py
@@ -192,6 +192,7 @@ class Layer(KeymapProvider, ABC):
         self._thumbnail_shape = (32, 32, 4)
         self._thumbnail = np.zeros(self._thumbnail_shape, dtype=np.uint8)
         self._update_properties = True
+        self._allow_thumbnail_update = True
         self._name = ''
         self.events = EmitterGroup(
             source=self,
@@ -637,6 +638,12 @@ class Layer(KeymapProvider, ABC):
         self._update_properties = False
         yield
         self._update_properties = True
+
+    @contextmanager
+    def block_thumbnail_update(self):
+        self._allow_thumbnail_update = False
+        yield
+        self._allow_thumbnail_update = True
 
     def _set_highlight(self, force=False):
         """Render layer highlights when appropriate.