--- a/napari/layers/shapes/_shapes_mouse_bindings.py
+++ b/napari/layers/shapes/_shapes_mouse_bindings.py
@@ -35,17 +35,19 @@ def select(layer, event):
         else:
             layer.selected_data = set()
     layer._set_highlight()
-    if len(layer.data) > layer._thumbnail_update_thresh:
-        update_thumbnail = False
-    else:
-        update_thumbnail = True
+
+    # we don't update the thumbnail unless a shape has been moved
+    update_thumbnail = False
     yield
 
     # on move
     while event.type == 'mouse_move':
         # Drag any selected shapes
         layer._move(layer.displayed_coordinates)
-        update_thumbnail = True
+
+        # if a shape is being moved, update the thumbnail
+        if layer._is_moving:
+            update_thumbnail = True
         yield
 
     # on release
@@ -58,9 +60,8 @@ def select(layer, event):
     elif layer._is_selecting:
         layer.selected_data = layer._data_view.shapes_in_box(layer._drag_box)
         layer._is_selecting = False
-        if len(layer.data) > layer._thumbnail_update_thresh:
-            update_thumbnail = False
         layer._set_highlight()
+
     layer._is_moving = False
     layer._drag_start = None
     layer._drag_box = None