--- a/napari/layers/shapes/shapes.py
+++ b/napari/layers/shapes/shapes.py
@@ -605,7 +605,7 @@ class Shapes(Layer):
         if self._update_properties:
             for i in self.selected_data:
                 self._data_view.update_edge_color(i, self._current_edge_color)
-        self.events.edge_color()
+            self.events.edge_color()
         self.events.current_edge_color()
 
     @property
@@ -620,7 +620,7 @@ class Shapes(Layer):
         if self._update_properties:
             for i in self.selected_data:
                 self._data_view.update_face_color(i, self._current_face_color)
-        self.events.face_color()
+            self.events.face_color()
         self.events.current_face_color()
 
     @property