--- a/napari/layers/shapes/shapes.py
+++ b/napari/layers/shapes/shapes.py
@@ -1589,7 +1589,7 @@ class Shapes(Layer):
                 edge_color=edge_color,
                 face_color=face_color,
                 z_index=z_index,
-                reindex_z=False,
+                z_refresh=False,
             )
             self._data_view._update_z_order()
             self.refresh_colors()
@@ -1603,7 +1603,7 @@ class Shapes(Layer):
         edge_color=None,
         face_color=None,
         z_index=None,
-        reindex_z=True,
+        z_refresh=True,
     ):
         """Add shapes to the data view.
 
@@ -1642,6 +1642,12 @@ class Shapes(Layer):
             same length as the length of `data` and each element will be
             applied to each shape otherwise the same value will be used for all
             shapes.
+        z_refresh : bool
+            If set to true, the mesh elements are reindexed with the new z order.
+            When shape_index is provided, z_refresh will be overwritten to false,
+            as the z indices will not change.
+            When adding a batch of shapes, set to false  and then call
+            ShapesList._update_z_order() once at the end.
         """
         if edge_width is None:
             edge_width = self.current_edge_width
@@ -1704,7 +1710,7 @@ class Shapes(Layer):
 
                 # Add shape
                 self._data_view.add(
-                    shape, edge_color=ec, face_color=fc, reindex_z=reindex_z
+                    shape, edge_color=ec, face_color=fc, z_refresh=z_refresh
                 )
 
         self._display_order_stored = copy(self.dims.order)