# Description
This PR implements some small changes to improve the responsiveness of the Shapes layer with a large number of shapes. In particular, based on the profiling we did in #1562 , I aimed to do the following:

- Reduce startup time
    - Delay re-indexing the mesh by z order until all shapes are added
    - Eliminate a redundant `Shapes._update_thumbnail()` call
- Decrease latency when switching interaction modes and selecting shapes
    - Eliminate a redundant `Shapes._update_thumbnail()` call
    - Do not call `Shapes._update_thumbnail()` when selecting shapes and there are more than 100 shapes
- Decrease latency during dragging
    - Delay updating the thumbnail until the drag is complete

### Profiling
I used the airspeed velocity benchmarks listed below (`/napari/benchmarks/benchmarks_shapes_layer.py`).

| Benchmark                                | n_shapes | Master [sec.] | PR [sec.]  | Change |
|------------------------------------------|----------|--------|------|--------|
| Shapes2DSuite.time_create_layer          | 256      | 1.93   | 1.25 | -35%   |
| ShapesInteractionSuite.time_select_shape | 256      | 0.79   | 0.01 | -99%   |
| ShapesInteractionSuite.time_drag_shape   | 256      | 1.23   | 0.18 | -85%   |

A couple of notes:
- `ShapesInteractionSuite.time_drag_shape` benchmark performs 5 drag events, the time per drag event is ~30 ms
- These only include the time for calculations performed on the model (i.e., they don't include the vispy draw time)
-  The layer creation is likely limited by the creation of each `Shape` object and the thumbnail generation.

### Demo
If you'd like to try it out, you can use the script below to generate a Shapes layer with 1000 shapes and compare between the PR and master. At least on my computer, this seems to be an improvement, but there are definitely more things we can do (e.g., speed up the calculation for selecting shapes with a bounding box)).

<details><summary>Demo script</summary>
<p>

```python
import numpy as np
import napari
from skimage import data

# make 3D image
blobs = data.binary_blobs(
    length=128, blob_size_fraction=0.05, n_dim=3, volume_fraction=0.1
).astype(float)


def make_square(lower_left_corner, width):
    upper_left_corner = lower_left_corner + [0, width]
    upper_right_corner = lower_left_corner + [width, width]
    lower_right_corner = lower_left_corner + [width, 0]

    return np.vstack((lower_left_corner, upper_left_corner, upper_right_corner, lower_right_corner))

# create random rectangles
n_shapes = 1000
np.random.seed(0)
corners = 100 * np.random.random((n_shapes, 2))
widths = 15 * np.random.random((n_shapes,))
squares = [make_square(llc, w) for llc, w in zip(corners, widths)]

base_cols = ['red', 'green', 'blue', 'white', 'yellow', 'magenta', 'cyan']
colors = np.random.choice(base_cols, size=n_shapes)

with napari.gui_qt():

    viewer = napari.view_image(blobs)

    shapes_layer = viewer.add_shapes(
        squares,
        shape_type='rectangle',
        face_color=colors,
        name='sliced',
    )
```

</p>
</details>

## Type of change
- [x] Bug-fix (non-breaking change which fixes an issue)

# References


# How has this been tested?
- [ ] all tests pass with my change

## Final checklist:
- [ ] My PR is the minimum possible work for the desired functionality
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my feature works