@kevinyamauchi 2020-08-20
Okay, I think this is ready for an initial round of review. I have added some details on what was changed and the profiling results above. I think there are still many things we can do, but I think these are the low hanging fruit and it is a step in the right direction.