@sofroniewn 2020-08-23
Amazing @kevinyamauchi 99% timesavings for dragging - this PR is huge. Thanks for including those benchmark stats!!!

I've left one final comment that we might want to additionally actually cap all thumbnail updates to at most use 100 random shapes if there are more than 100 shapes on that slice (note not 100 overall, but on that slice). For a 32 x 32 pixel thumbnail, I wonder if more than 100 are ever needed. You could also take the top 100 shapes according to z-index. You could also then remove the additional split on if to block thumbnail during the changing of mode, which should probably never need a thumbnail calculation.

Let me know if you want to make those two changes, I think it is very close to merge.