@kevinyamauchi 2020-08-24
@sofroniewn, I have done the following, which I think addresses your comments.

- Made it such that the thumbnail is not updated for any layer mode change
- Made it so that the selection events never update the thumbnail. The thumbnail still updates if a shape is moved, modified, or created.
- Made it so when the number of shapes _in view_ exceeds a threshold (`Shapes._max_shapes_thumbnail`), a number of in view shapes are randomly selected to be rasterized. I set the threshold to 100 for now, but we can relax this as we further improve the rasterization code (this yields update times of ~100 ms on my Macbook Air).

I have updated the benchmarks in this comment: https://github.com/napari/napari/pull/1561#issuecomment-678771901