@sofroniewn 2020-08-24
Hi @kevinyamauchi thanks for all those updates. Playing with the new behavior I have one final request which is to not make the selection randomly, but take the top `Shapes._max_shapes_thumbnail` shapes for the thumbnail, should be a simple one line change, but you should make sure to take the top 100 not the bottom 100.

Reason for this is shown in the gif placed in the inline comment, but right now a small change to the shapes triggers a big change to the thumbnail as the random 100 chosen are different every time which is quite visually distracting if you're moving shapes around.

Otherwise everything looks great, and we can merge after this!