@kevinyamauchi 2020-08-25
Thanks for the careful review, @sofroniewn ! I agree that taking the "top" (by z order) `max_shapes` is the way to go so the thumbnail is stable for dense layers. I have made the suggested change.